requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
motor==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-cloud-texttospeech==2.14.1
google-api-python-client==2.86.0
google-generativeai==0.3.1
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
Pillow==9.5.0
pytesseract==0.3.10
pytest==7.3.1
pytest-asyncio==0.21.0
aiohttp==3.8.4